+++ src/main/kotlin/icu/windea/pls/ep/index/ParadoxIndexInfoSupports.Core.kt
@@ -4,8 +4,10 @@ import com.intellij.codeInsight.highlighting.*
 import com.intellij.psi.*
 import com.intellij.psi.util.*
 import icu.windea.pls.core.*
+import icu.windea.pls.lang.PlsManager
 import icu.windea.pls.lang.psi.*
 import icu.windea.pls.lang.util.*
+import icu.windea.pls.lang.withState
 import icu.windea.pls.localisation.psi.*
 import icu.windea.pls.model.*
 import icu.windea.pls.model.constraints.*
@@ -64,7 +66,7 @@ class ParadoxDynamicValueIndexInfoSupport : ParadoxIndexInfoSupport<ParadoxDynam
         }
         references.forEach f@{ reference ->
             if (!constraint.canResolve(reference)) return@f
-            val resolved = reference.resolve()
+            val resolved = withState(PlsManager.resolveForMergedIndex) { reference.resolve() }
             if (resolved !is ParadoxDynamicValueElement) return@f
             resolved.dynamicValueTypes.forEach { dynamicValueType ->
                 val info = ParadoxDynamicValueIndexInfo(resolved.name, dynamicValueType, resolved.readWriteAccess, resolved.parent.startOffset, resolved.gameType)
@@ -80,7 +82,7 @@ class ParadoxDynamicValueIndexInfoSupport : ParadoxIndexInfoSupport<ParadoxDynam
         val references = ParadoxExpressionManager.getExpressionReferences(element)
         references.forEach f@{ reference ->
             if (!constraint.canResolve(reference)) return@f
-            val resolved = reference.resolve()
+            val resolved = withState(PlsManager.resolveForMergedIndex) { reference.resolve() }
             if (resolved !is ParadoxDynamicValueElement) return@f
             resolved.dynamicValueTypes.forEach { dynamicValueType ->
                 val info = ParadoxDynamicValueIndexInfo(resolved.name, dynamicValueType, resolved.readWriteAccess, resolved.parent.startOffset, resolved.gameType)
@@ -126,7 +128,7 @@ class ParadoxParameterIndexInfoSupport : ParadoxIndexInfoSupport<ParadoxParamete
         }
         references.forEach f@{ reference ->
             if (!constraint.canResolve(reference)) return@f
-            val resolved = reference.resolve()
+            val resolved = withState(PlsManager.resolveForMergedIndex) { reference.resolve() }
             if (resolved !is ParadoxParameterElement) return@f
             //note that element.startOffset may not equal to actual parameterElement.startOffset (e.g. in a script value expression)
             val info = ParadoxParameterIndexInfo(resolved.name, resolved.contextKey, resolved.readWriteAccess, element.startOffset, resolved.gameType)
@@ -171,7 +173,7 @@ class ParadoxLocalisationParameterIndexInfoSupport : ParadoxIndexInfoSupport<Par
         }
         references.forEach f@{ reference ->
             if (!constraint.canResolve(reference)) return@f
-            val resolved = reference.resolve()
+            val resolved = withState(PlsManager.resolveForMergedIndex) { reference.resolve() }
             if (resolved !is ParadoxLocalisationParameterElement) return@f
             val info = ParadoxLocalisationParameterIndexInfo(resolved.name, resolved.localisationName, element.startOffset, resolved.gameType)
             addToFileData(info, fileData)