+++ src/main/kotlin/icu/windea/pls/ep/scope/ParadoxDefinitionInferredScopeContextProviders.kt
@@ -87,7 +87,7 @@ class ParadoxBaseDefinitionInferredScopeContextProvider : ParadoxDefinitionInfer
         return withRecursionGuard {
             withRecursionCheck("${definitionInfo.name}:${definitionInfo.type}") {
                 val indexId = ParadoxIndexInfoType.InferredScopeContextAwareDefinition
-                ParadoxIndexManager.processQueryForIndexInfo(ParadoxScriptFileType, indexId, project, gameType, searchScope) p@{ file, infos ->
+                ParadoxIndexManager.processQueryForMergedIndex(ParadoxScriptFileType, indexId, project, gameType, searchScope) p@{ file, infos ->
                     val psiFile = file.toPsiFile(project) ?: return@p true
                     infos.forEach f@{ info ->
                         ProgressManager.checkCanceled()
@@ -194,7 +194,7 @@ class ParadoxEventInOnActionInferredScopeContextProvider : ParadoxDefinitionInfe
             if (depth == 1) stackTrace.addLast(thisEventName)
 
             val indexId = ParadoxIndexInfoType.EventInOnAction
-            ParadoxIndexManager.processQueryForIndexInfo(ParadoxScriptFileType, indexId, project, gameType, searchScope) p@{ file, infos ->
+            ParadoxIndexManager.processQueryForMergedIndex(ParadoxScriptFileType, indexId, project, gameType, searchScope) p@{ file, infos ->
                 val psiFile = file.toPsiFile(project) ?: return@p true
                 infos.forEach f@{ info ->
                     ProgressManager.checkCanceled()
@@ -302,7 +302,7 @@ class ParadoxEventInEventInferredScopeContextProvider : ParadoxDefinitionInferre
 
             val toRef = "from".repeat(depth)
             val indexId = ParadoxIndexInfoType.EventInEvent
-            ParadoxIndexManager.processQueryForIndexInfo(ParadoxScriptFileType, indexId, project, gameType, searchScope) p@{ file, infos ->
+            ParadoxIndexManager.processQueryForMergedIndex(ParadoxScriptFileType, indexId, project, gameType, searchScope) p@{ file, infos ->
                 val psiFile = file.toPsiFile(project) ?: return@p true
                 infos.forEach f@{ info ->
                     ProgressManager.checkCanceled()
@@ -454,7 +454,7 @@ class ParadoxOnActionInEventInferredScopeContextProvider : ParadoxDefinitionInfe
 
             val toRef = "from".repeat(depth)
             val indexId = ParadoxIndexInfoType.OnActionInEvent
-            ParadoxIndexManager.processQueryForIndexInfo(ParadoxScriptFileType, indexId, project, gameType, searchScope) p@{ file, infos ->
+            ParadoxIndexManager.processQueryForMergedIndex(ParadoxScriptFileType, indexId, project, gameType, searchScope) p@{ file, infos ->
                 val psiFile = file.toPsiFile(project) ?: return@p true
                 infos.forEach f@{ info ->
                     ProgressManager.checkCanceled()