+++ src/main/kotlin/icu/windea/pls/lang/index/ParadoxDefineIndex.kt
@@ -16,7 +16,7 @@ import java.io.*
  */
 class ParadoxDefineIndex : ParadoxFileBasedIndex<Map<String, ParadoxDefineIndexInfo.Compact>>() {
     companion object {
-        private const val VERSION = 62 //1.4.0
+        private const val VERSION = 64 //1.4.0
     }
 
     override fun getName() = ParadoxIndexManager.DefineName