+++ src/main/kotlin/icu/windea/pls/lang/index/ParadoxFileBasedIndex.kt
@@ -7,48 +7,40 @@ import com.intellij.util.gist.*
 import com.intellij.util.indexing.*
 import com.intellij.util.io.*
 import icu.windea.pls.core.*
-import icu.windea.pls.lang.*
 import java.io.*
 
-abstract class ParadoxFileBasedIndex<T : Any> : SingleEntryFileBasedIndexExtension<Map<String, T>>() {
+abstract class ParadoxFileBasedIndex<T> : FileBasedIndexExtension<String, T>() {
     abstract override fun getVersion(): Int
 
-    override fun getIndexer(): SingleEntryIndexer<Map<String, T>> {
-        return object : SingleEntryIndexer<Map<String, T>>(false) {
-            override fun computeValue(inputData: FileContent): Map<String, T> {
-                val psiFile = inputData.psiFile
-                return buildFileData(psiFile)
-            }
+    override fun getIndexer(): DataIndexer<String, T, FileContent> {
+        return DataIndexer { inputData ->
+            val psiFile = inputData.psiFile
+            buildFileData(psiFile)
         }
     }
 
-    override fun getValueExternalizer(): DataExternalizer<Map<String, T>> {
-        return object : DataExternalizer<Map<String, T>> {
-            override fun save(storage: DataOutput, value: Map<String, T>) {
-                storage.writeIntFast(value.size)
-                if (value.isEmpty()) return
-                value.forEach { (k, v) ->
-                    storage.writeUTFFast(k)
-                    writeData(storage, v)
-                }
+    override fun getKeyDescriptor(): KeyDescriptor<String> {
+        return EnumeratorStringDescriptor.INSTANCE
+    }
+
+    override fun getValueExternalizer(): DataExternalizer<T> {
+        return object : DataExternalizer<T> {
+            override fun save(storage: DataOutput, value: T) {
+                writeData(storage, value)
             }
 
-            override fun read(storage: DataInput): Map<String, T> {
-                val size = storage.readIntFast()
-                if (size == 0) return emptyMap()
-                return buildMap {
-                    repeat(size) {
-                        val key = storage.readUTFFast()
-                        val value = readData(storage)
-                        put(key, value)
-                    }
-                }
+            override fun read(storage: DataInput): T {
+                return readData(storage)
             }
         }
     }
 
     override fun getInputFilter(): FileBasedIndex.InputFilter {
         return FileBasedIndex.InputFilter { file -> filterFile(file) && !useLazyIndex(file) }
+    }
+
+    override fun dependsOnFileContent(): Boolean {
+        return true
     }
 
     private val indexName by lazy {
@@ -90,12 +82,7 @@ abstract class ParadoxFileBasedIndex<T : Any> : SingleEntryFileBasedIndexExtensi
 
     private fun buildFileData(file: PsiFile): Map<String, T> {
         return buildMap {
-            try {
-                PlsManager.indexing.set(true)
-                indexData(file, this)
-            } finally {
-                PlsManager.indexing.remove()
-            }
+            indexData(file, this)
         }
     }