+++ src/main/kotlin/icu/windea/pls/lang/index/ParadoxFileLocaleIndex.kt
@@ -14,7 +14,7 @@ import java.util.*
  */
 class ParadoxFileLocaleIndex : ScalarIndexExtension<String>() {
     companion object {
-        private const val VERSION = 62 //1.4.0
+        private const val VERSION = 64 //1.4.0
     }
 
     override fun getName() = ParadoxIndexManager.FileLocaleName