+++ src/main/kotlin/icu/windea/pls/lang/search/ParadoxGlobalScriptedVariableSearcher.kt
@@ -6,6 +6,7 @@ import com.intellij.openapi.project.*
 import com.intellij.psi.search.*
 import com.intellij.util.*
 import icu.windea.pls.core.*
+import icu.windea.pls.lang.PlsManager
 import icu.windea.pls.lang.index.ParadoxIndexManager
 import icu.windea.pls.lang.search.scope.*
 import icu.windea.pls.script.psi.*
@@ -17,6 +18,9 @@ import icu.windea.pls.script.psi.*
  */
 class ParadoxGlobalScriptedVariableSearcher : QueryExecutorBase<ParadoxScriptScriptedVariable, ParadoxGlobalScriptedVariableSearch.SearchParameters>() {
     override fun processQuery(queryParameters: ParadoxGlobalScriptedVariableSearch.SearchParameters, consumer: Processor<in ParadoxScriptScriptedVariable>) {
+        //#141 如果正在为 ParadoxMergedIndex 编制索引并且正在解析引用，则直接跳过
+        if(PlsManager.resolveForMergedIndex.get() == true) return
+
         ProgressManager.checkCanceled()
         if(queryParameters.project.isDefault) return
         val scope = queryParameters.selector.scope