+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## PLANNED
 
 * [ ] 优化：为本地化文本中的属性引用&命令的传入参数提供关于其中使用的格式标签的说明（文档注释、枚举、工具类等）
+* [ ] 优化：优化默认生成的一些PsiElement的PSI访问方法，如果必要，从后往前遍历子节点
 * [ ] 提供代码检查，用于高亮本地化文本中可能的格式错误（例如多余的`!`） / Provide the code inspection to highlight possible format errors in loc text (such as redundant `!`) 
 * [ ] 提供对魔法注释的初步支持 / Provide initial support for magic Comments
 * [ ] 提供用于生成并修补脚本的魔法注释，实现相关功能 / Provide magic comments to generate and patch scripts, implement related features
@@ -15,7 +16,7 @@
 * [X] #94 优化CWT文件解析器的性能 / Optimize performance for CWT file parser
 * [ ] #137 [VIC3/CK3] Support special localizations - Basic support
 * [X] #140 修复与本地化命令连接相关的一些问题（规则解析、代码导航等） / Fixed some problems about localisation command links (Config resolving, code navigation, etc.)
-* [ ] #141 [Stellaris] Indexing process should not rely on non-indexed file data
+* [X] #141 [Stellaris] Indexing process should not rely on non-indexed file data
 * [X] 通过懒解析CWT文件中的选项注释来尝试优化性能 / Try to optimize performance by lazily parsing option comments in CWT files
 * [X] 通过懒解析本地化文件中的本地化文本来尝试优化性能与提高代码灵活性 / Try to optimize performance and improve code flexibility by lazily parsing localisation text in localisation files
 * [X] 修复：本地化图标如果对应一个sprite，无法正常适用用法高亮 / FIX: Localisation icons could not be properly highlighted if it will be resolved to a sprite